Change MDI sub-window events to not use signals
Previously MDI sub-window events were overridden to emit signals.
Its "parent" MDI area would then connect to them and process them.
With this, the sub-windows are free to not know about the MDI area.

This led to a potential race condition when sub-windows were closed.
The sub-windows are set to `WA_DeleteOnClose` which means
the default close event handler would `deleteLater` the sub-window,
and then hide the window.
Note that no explicit removal from the QObject tree occurs,
as deletion would automatically detach the sub-window from the MDI area.
In theory, this deletion may occur at any time.
In particular, it can happen before the hiding or signals are emitted.
In such cases,
the hide event would not have access to the MDI area via `mdiArea()`
or the deletion disconnects signals
and the MDI area would not have an opportunity to handle the events.

Also, the original premise that the sub-windows
need not know about their parents was not sound
since emitting singals already required overidding.
If that has to occur anyway, then the MIDarea slot functions
can be directly called from in the sub-window event handlers instead,
and that is what is used.
Unlike the signal handling method, this is much more deterministic.

This also liberally adds a few debug logging statements
that helped in tracking down this race condition.
Note also that no unit tests were added to mimic this race condition.
I was not able to trigger it in unit tests, only in GUI.