Refactor out EventSetter from tests
The class `EventSetter` simply `set`-s a `threading.Event`
inside a `dispatch` method, which is used in `watchdog` event handlers.
It is useful for checking whether a particular file `watch`
has `dispatch`-ed events to all handlers.
In particular, by ensuring `EventSetter` is added last as a handler,
current implementation detail allows us to assume that
all other handlers for a particular `watch` had been `dispatch`-ed to,
if the `threading.Event` becomes `set`.

This class is being used in more than one unit test files.
So it is refactored out.