Refactor out `Notification` from phile-notify-cli
The notify CLI and GUI use the same logic
in determining notification paths and how to manipulate them.
The GUI code already created a `Notification` interface
with the intention of sharing it with the CLI code.
This refactoring removes the duplication from the path handling,
by using `Notification` in the CLI code as well.

This moves `Notification`, `Configuration` and `default_paths`
into the same module as they are used together.