Remove exceptions from watchdog_wrapper start/stop
Behaviour of `Observer` should extend `watchdog.observers.Observer`.
The current`Observer.stop` throws exceptions
when the `Observer` was not started or already `stop`-ped,
which is different from the parent behaviour.
This is undesirable for an extension.
So the exceptions will now not be raised.

The `FileSystemMonitor` and `FileSystemSignalEmitter`
provides a way to integrate `Observer` API into `PySide2`.
Unlike `Observer`, their API should be consistent
with `PySide2` behaviours, in particular,
their `start`/`stop` methods mimic `show`/`hide` behaviours.
The latter pair does nothing if the target was already shown or hidden.
And so the monitors and emitters should also do that
if they were already started or stopped.
The current behaviour is that they raise an exception in those cases.
They will be changed to not do anything,
essentially an idempotent behaviour.

There is a slight difference from the `show`/`hide` pair
in the sense that a stopped monitor cannot be started again,
because a thread cannot be started twice.
In that case, an exception is appropriate,
since a thread cannot be `start`-ed twice.

The purose of `Observer` is clarified a little more in its docstring.