Add documentation build step in tox
Using the `sphinx` module for building documentation.
Using the `recommonmark` module for Markdown parsing in documentations.
Using `===` instead of hash for Markdown title in `README.md`,
because `recommonmark` is currently unable to parse hash titles.
No documentation is set up to be built yet
other than the `README.md` as content.