Fix unhandled exception from parsing tray file
Tray files represent items to display in the system tray.
They are read and parsed for information on what to display.
It is possible for the tray files to be modified while the read occurs,
leading to an exception being raised.

The exceptions are now handled.
They are ignored after logging a warning.
There is little we can do from the reading side.

Tray file parsing now also caches the entire file before parsing,
to reduce the chance of the file being changed before reading completes.

Adds unit tests for loading bad tray files.

Fixes: https://github.com/BoniLindsley/phile/issues/1