test: Fix race condition in unit tests
-   Replaces all uses of `EventSetter` by the new `ThreadedMock`.
    The new mock subclass allows for `assert_called_soon`
    wrapping common uses of waiting `EventSetter.set` to be called.
    Also provides `assert_called_with_soon`
    to also check for arguments provided.

Details
-------

Some unit tests check that file system events are handled as expected.
The events are handled in a worker thread.
The check is done by installing another event handler
to respond to the events in the worker thread.
after handlers under-test in worker threads are installed.
It was assumed that handler would be dispatched
in the order they were installed.
By relying on this order, by the time the extra handler is dispatched,
the worker thread handlers could be assumed to have been dispatched.
The unit tests can check for post-condition of the worker handlers,
by waiting for the extra installed handlers to be called.

Unfortunately, the order of handler dispatch is not guaranteed,
as they are stored in a set in the `watchdog` module.
In other words, the unit tests work by luck so far
and it just so happens that the sets iterate the handlers
in the necessary order.

This order dependency is now removed.
Also moves some common loops for post-condition checks
into assertion methods.

Fixes: https://github.com/BoniLindsley/phile/issues/2