Use FileSystemMonitor not Observer in GUI classes
GUI classes, such as `MainWindow` and `GuiIconList`,
use watchdog `Observer`-s to monitor for file changes.
The `Observer`-s dispatch events and call handlers in worker threads
to allow the GUI classes to respond to file changes.
However, these event handlers cannot call GUI functions,
because GUI functions are typically not thread-safe.
So the GUI classes use the `Observer`-s
by wrapping them in `FileSystemMonitor`-s to manage their lifetimes,
and `FileSystemSignalEmitter`
which translate the watchdog events into GUI signals
in order to handle them inside the GUI event loop.

The GUI classes did indeed take an `Observer` instance in its parameter.
Since `FileSystemMonitor` takes ownership of the given `Observer`,
it means the GUI classes do as well,
and so sharing the same observer with multple GUI classes
is not supported.

The GUI classes now take a `FileSystemMonitor` argument
instead of `Observer` to allow for sharing with other GUI classes.
This is in preparation for displaying multiple such GUI classes
in the same application.

Fixes: https://github.com/BoniLindsley/phile/issues/6