test: Use `addCleanup` instead of `tearDown`
Unit test `tearDown` methods run if their corresponding `setUp` succeeds.
This poses a problem for PySide2 `QCoreApplication`
that must be single-instance.
For example, if a previous `setUp` or `tearDown`
was unable to destroy the `QCoreApplication` it uses,
then all future unit tests will be able to create one.
This has a cascading effect of failing all such future tests.

Using `addCleanup` reduces the chance of `QCoreApplication`
not being cleaned up properly,
as it inserts the clean-up code that will be ran
even if `setUp` fails.

This has a nice side-effect of create-and-forget,
since create and clean-up code are right next to each other,
reducing maintenance burden by a bit.