test: Remove `ThreadedMock` side-effect
The class had taken over `side-effect`
to call a callable given in a `target` parameter.
This was done because I was not aware
of the `wraps` parameter from `unittest.mock.Mock`
that achieve the same thing.
So now, we just forward `__call__` to parent,
and add any required processing.
This conforms to user expectation of how to use the class
since it is now the same as the `Mock` class.