fix: Kill not wait in `ControlMode` finaliser
When the finaliser is called, the client state may not be stable.
Waiting would make the finaliser take longer than it should.
This can make shutdown time perceived by user much longer.

Instead, the tmux client process is killed in the finaliser
if it had not already been terminated.
This leaves the responsibility of proper clean-up to user code,
while ensuring the tmux client process is properly terminated.