refactor: Watchdog event handling in GUI
Watchdog event handling in GUI was done
by using a callable class that specially posts Watchdog events
into the Qt event loop to be handled.
This is necessary to allow the events to be handled in a thread-safe way
since GUI functions are typically not thread-safe.
Once the posted event was received in the Qt loop,
a Qt signal would then be emitted.
By connecting to the signal,
anyone interested in handling the event could do so.

The callable class only accepted a Watchdog events
because the signal emitte required a fixed signature.
By dropping the signal handling feature, which was not used,
the callable can now accept arbitrary arguments.
The class is renamed to `CallSoon`
mimicking `call_soon` in `asyncio` event loops.