refactor: add method converting events to paths
Events from the watchdog module indicates what is changed to paths.
But due to the race condition
between the event being emitted and processed,
event handlers cannot rely on the changes given in the events
and have to operate directly on the paths.
So a conversion is needed in all handler code.