refactor: generalise sorter for notify
The sorter functioned as a cache for known notify files,
and sorts them in order for displaying.
So it is renamed to `SortedLoadCache`,
used `typing.Protocol` to generalise its interface.
It is generalised because the same cache functionality is used
for tray file sorting.
So it is in preparation for that use.