refactor: phile.notify.gui to use SortedLoadCache
It was previously a big block of code for loading.
Now it is split into subroutines.
There are still more refactoring to do -- one step at a time.

Also removes `__hash__` from `phile.notify.File`.
It was only used in phile.notify.gui for use in a dictionary
which has now been refactored out.t