refactor: remove `phile.tray.event` module
It is not used anymore.
It had contained `Filter`, `Converter` and `Sorter` classes.
The first two simply check path parent and suffix,
and create a `phile.tray.File` respectively, when given a path.
Creating new callables to achieve these does not make sense.
It is better to make callables
that are easy to wrap in `functools.partial` instead.
They are merged into `phile.tray.File`
as `check_path` and `from_path_stem` methods.

The last is responsible of keeping track of known tray files.
It is replaced by `phile.data.SortedLoadCache` which does the same thing.
In order to use the cache object,
some methods in other classes have their signatures modified
to fit with the cache object interface.
The `load` method of the `phile.tray.File` class
is modified to not raise known exceptions
and return False for those errors.