test: fix gui closing not being detected
It depended on detecting whether triggers had been processed
before checking whether the GUI is closed.
But the detection only checks whether some trigger was processed
rather than whether a close trigger was processed.
This led to a race condition between the detection and checking.

This is fixed by checking specifically for the close trigger file.