test: fix notify_monitor worker thread shutdown
There was a race condition between checking that a worker thread is alive
and closing the event loop active in the worker thread.
If the event loop is closed in between the check and close
an exception would be raised from the failing to close a second time.

In accordance to EAFP, the event loop will be closed instead
(by cancelling the task in progress)
and ignoring the expected exception from it.