refactor: rename `datetime_display` to `datetime`
The word "display" does not add much
to indicate what `tray.publishers.datetime` does.

Prepends the weekday numeric value in its text display with a "w"
to hopefully suggest what it means.
For example, "w4" for Thursday.
Otherwise, it is just a numeric value on its own
in the middle of date and time values.