refactor: trigger handlers are now nullary
Callbacks are called in response to trigger files being deleted.
The callbacks are given the name of the trigger files being deleted.
However, none of the callbacks so far need to know the trigger name,
and they are all wrapped in a lambda to remove the argument.
So it makes sense to simplify the handler interface
to not take any arguments,