refactor: Environ backup class into phile.os
The class used to be defined in a PySide2 test module.
But it is also used in a tmux test module.
This pulls the class into its own module
so that the tmux test module does not have to depend on the PySide2 one.
Also adds a unit test for the class.