fix: possible syntax error in `ResponseLine`
A trailing comma in the argument list of `typing.Union`
might have caused Python to interpret the argument list as a single list
instead of arguments to the `Union`,
and a list would not make sense as an argument to `Union`.

This fix bypasses this issue by not defining the `Union` at all.