refactor: use watchdog asyncio event view
Observers now return a view instead of a queue on `schedule` call.
Since a view returns events
starting from the "current" state of the queue,
whereas the "current" state of a queue may change at any time,
callers of `schedule` does not need to call `__aiter__` to obtain a view
to not miss events.